split Go sources and add aggregate html report
